Finished the interface comments, removed a useless synchronization from the Clock class
