--- src/main/java/it/tristana/commons/interfaces/util/Status.java
@@ -1,10 +1,0 @@
-package it.tristana.commons.interfaces.util;
-
-public enum Status {
-
-	WAITING,
-	STARTING,
-	PLAYING,
-	ENDING,
-	DISABLED
-}
